Introduce HoverTableCellView to handle the hover mouse in many AutoCompleteCells (mac)
