+++ src/ui/osx/TogglDesktop/Feature/TimeEntryEditor/Workspace/WorkspaceCellView.swift
@@ -8,55 +8,17 @@
 
 import Cocoa
 
-final class WorkspaceCellView: NSTableCellView {
+final class WorkspaceCellView: HoverTableCellView {
 
     static let cellHeight: CGFloat = 34.0
 
     // MARK: OUTLET
 
-    @IBOutlet weak var contentContainerView: NSBox!
     @IBOutlet weak var titleTextField: NSTextField!
 
     // MARK: Public
 
-    override func awakeFromNib() {
-        super.awakeFromNib()
-
-        initCommon()
-        initTracking()
-    }
-
-    override func prepareForReuse() {
-        super.prepareForReuse()
-        NSCursor.arrow.set()
-        contentContainerView.animator().alphaValue = 0.0
-    }
-
     func render(_ item: Workspace) {
         titleTextField.stringValue = item.name
-    }
-
-    override func mouseExited(with event: NSEvent) {
-        super.mouseExited(with: event)
-        NSCursor.arrow.set()
-        contentContainerView.animator().alphaValue = 0.0
-    }
-
-    override func mouseEntered(with event: NSEvent) {
-        super.mouseEntered(with: event)
-        NSCursor.pointingHand.set()
-        contentContainerView.animator().alphaValue = 1.0
-    }
-
-    fileprivate func initCommon() {
-        contentContainerView.alphaValue = 0
-    }
-
-    fileprivate func initTracking() {
-        let trackingArea = NSTrackingArea(rect: bounds,
-                                          options: [.activeAlways, .inVisibleRect, .mouseEnteredAndExited],
-                                          owner: self,
-                                          userInfo: nil)
-        addTrackingArea(trackingArea)
     }
 }
